Implement grpc UIInput and UIOutput
We can use the grpc streaming features to implement the io for the
Provider.Input and Provisioner.Apply calls.

The ProviderClient Input operation is a bi-directional stream, first
sending a config, and then user replies to the server. Responses without a
config are forwarded to the client's UIInput instance, and a response
with a config is the end of the request.

The ProvisionerClient apply has a streaming response. A single request
is sent to the server, and any response strings are forwarded to the
client's IUOutput instance.